+++ npm-api-maker/src/link.jsx
@@ -10,6 +10,8 @@ import memo from "set-state-compare/build/memo.js"
 
 export default memo(shapeComponent(class ApiMakerLink extends BaseComponent {
   static propTypes = {
+    paddingHorizontal: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    paddingVertical: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     usePressable: PropTypes.bool
   }
 
@@ -18,7 +20,8 @@ export default memo(shapeComponent(class ApiMakerLink extends BaseComponent {
   }
 
   render() {
-    const {dataSet, to, onClick, onPress, testID, usePressable, ...restProps} = this.props
+    const {dataSet, to, onClick, onPress, paddingHorizontal, paddingVertical, style, testID, usePressable, ...restProps} = this.props
+    const linkStyle = this.linkStyle({paddingHorizontal, paddingVertical, style})
 
     if (Platform.OS == "web" && !usePressable) {
       return (
@@ -26,14 +29,53 @@ export default memo(shapeComponent(class ApiMakerLink extends BaseComponent {
           {...dataSetToAttributes(Object.assign({testid: testID}, dataSet))} // eslint-disable-line prefer-object-spread
           href={to || "#"}
           onClick={this.tt.onLinkClicked}
+          style={linkStyle}
           {...restProps}
         />
       )
     }
 
     return (
-      <Pressable dataSet={dataSet} onPress={this.tt.onPress} testID={testID} {...restProps} />
+      <Pressable dataSet={dataSet} onPress={this.tt.onPress} style={linkStyle} testID={testID} {...restProps} />
     )
+  }
+
+  /** Returns a cached style object with padding props applied. */
+  linkStyle({paddingHorizontal, paddingVertical, style}) {
+    const stylePaddingHorizontal = style?.paddingHorizontal
+    const stylePaddingVertical = style?.paddingVertical
+    const paddingHorizontalDefined = paddingHorizontal !== null && paddingHorizontal !== undefined
+    const paddingVerticalDefined = paddingVertical !== null && paddingVertical !== undefined
+    const stylePaddingHorizontalDefined = stylePaddingHorizontal !== null && stylePaddingHorizontal !== undefined
+    const stylePaddingVerticalDefined = stylePaddingVertical !== null && stylePaddingVertical !== undefined
+    const hasPaddingHorizontal = paddingHorizontalDefined || stylePaddingHorizontalDefined
+    const hasPaddingVertical = paddingVerticalDefined || stylePaddingVerticalDefined
+
+    return this.cache("linkStyle", () => {
+      if (hasPaddingHorizontal || hasPaddingVertical) {
+        const styleObject = style || {}
+        const resolvedPaddingHorizontal = paddingHorizontalDefined ? paddingHorizontal : stylePaddingHorizontal
+        const resolvedPaddingVertical = paddingVerticalDefined ? paddingVertical : stylePaddingVertical
+        const nextStyle = {...styleObject}
+
+        delete nextStyle.paddingHorizontal
+        delete nextStyle.paddingVertical
+
+        if (resolvedPaddingHorizontal !== null && resolvedPaddingHorizontal !== undefined) {
+          if (nextStyle.paddingLeft === null || nextStyle.paddingLeft === undefined) nextStyle.paddingLeft = resolvedPaddingHorizontal
+          if (nextStyle.paddingRight === null || nextStyle.paddingRight === undefined) nextStyle.paddingRight = resolvedPaddingHorizontal
+        }
+
+        if (resolvedPaddingVertical !== null && resolvedPaddingVertical !== undefined) {
+          if (nextStyle.paddingTop === null || nextStyle.paddingTop === undefined) nextStyle.paddingTop = resolvedPaddingVertical
+          if (nextStyle.paddingBottom === null || nextStyle.paddingBottom === undefined) nextStyle.paddingBottom = resolvedPaddingVertical
+        }
+
+        return nextStyle
+      }
+
+      return style
+    }, [hasPaddingHorizontal, hasPaddingVertical, paddingHorizontal, paddingVertical, style, stylePaddingHorizontal, stylePaddingVertical])
   }
 
   onLinkClicked = (e, ...restArgs) => {